tooling: revert to React, React DOM & React scripts pre-React 17
While it would be great to update to the latest version of React and all
its tooling, there seems to be an issue with typescript versions,
react-scripts and the new JSX transform wreaking havoc on the JSX
checks. While the issue was supposedly fixed here
(https://github.com/facebook/create-react-app/pull/9869), upgrading to
react-scripts 4.0.1 doesn't fix the problem. Neither does upgrading AND
pinning typescript to 4.0.5 so it doesn't overwrite the tsconfig.json
file. Let's wait a few versions ahead for react-scripts so we can be
sure this is fixed.