Console: replace Apollo Client with React Query
The rationale for this change is mainly due to how heavy / unflexible
Apollo client can be when thinking outside a GraphQL context. As we
might be integrating with external APIs for DAO Activity / Snapshot, it
makes sense to move to something more protocol agnostic, but that still
provides all the nice things, such as retries and caching. react-query
has become a community standard and is well maintained, so, in my
opinion, is a good choice.