Console: base routing on network / chain ID rather than DAO address (#245)
* console: base routing on network / chain ID rather than DAO address

* console: move title inside button and remove alt prop for screen readers

Quoting Pierre:
"We should use an empty alt here, since we have the title right after it
(“Govern Console”). When a logo is being used in a header, it is
generally recommended to use either an empty alt (when the textual
version is present) or the name of the site / app directly, rather than
describing the image itself. This is to avoid screen readers repeating
“Grey Eagle Aragon logo − Govern Console” every time the page loads 😄

Now since we are adding a button here, we need a way to describe it. We
could use a title to describe the action, but should we move the h1
inside the button instead? I think it would make sense to make them both
clickable.
"