Console: remove the ChainID provider
The current ChainID is now only stored in the URL, so there is no need
to maintain the state inside a provider or in localStorage. The env var
(CHAIN_ID) is still used to determine the default.

useChainId() has been moved to lib/chain-id.ts, and a lib/routing.ts has
been created to centralize the different routes (for now it only
contains goHome()).

The redirection from / has been removed and replaced by a useEffect(),
so that we can redirect any other incorrect path too. This is probably
something we want to change back whenever we’ll have a global error.